+++ R/ModuleConnectivity.R
@@ -15,6 +15,7 @@
 #' @param slot Slot in seurat_obj, default to normalized 'data' slot.
 #' @param sparse logical indicating whether or not to run the correlation using a sparse matrix.
 #' @param reassign_modules logical indicating whether or not to reassign genes to different co-expression modules if their kME value in the assigned module is negative.
+#' @param TOM_use name of the hdWGCNA experiment that contains the TOM which will be used to compute the intramodular degree of each gene. If the TOM is not found, this calculation will be skipped.
 #' @param wgcna_name The name of the hdWGCNA experiment in the seurat_obj@misc slot
 #' @details
 #' ModuleConnectivity computes the eigengene-based connectivity (kME) of each feature
@@ -50,6 +51,7 @@ ModuleConnectivity <- function(
   layer = 'data',
   sparse = TRUE,
   reassign_modules = TRUE,
+  TOM_use = NULL,
   wgcna_name = NULL,
   ...
 ){
@@ -57,16 +59,27 @@ ModuleConnectivity <- function(
   # set as active assay if wgcna_name is not given
   if(is.null(wgcna_name)){wgcna_name <- seurat_obj@misc$active_wgcna}
   CheckWGCNAName(seurat_obj, wgcna_name)
-  
-  # get module df, wgcna genes, and wgcna params:
+  if(is.null(TOM_use)){TOM_use <- wgcna_name}
+
+  # get module df, wgcna genes, TOM, and wgcna params:
   modules <- GetModules(seurat_obj, wgcna_name)
   MEs <- GetMEs(seurat_obj, harmonized, wgcna_name)
   genes_use <- as.character(modules$gene_name)
   params <- GetWGCNAParams(seurat_obj, wgcna_name)
 
+  # try to load the TOM
+  TOM <- tryCatch({
+      GetTOM(seurat_obj, TOM_use)
+   },
+   error = function(e){
+    warning('TOM not found')
+   }
+  )
+
   # get the assay
   if(is.null(assay)){assay <- DefaultAssay(seurat_obj)}
-
+  
+  # select which cells to use for this analysis
   if(!is.null(group.by)){
     cells.use <- seurat_obj@meta.data %>% subset(get(group.by) %in% group_name) %>% rownames
     MEs <- MEs[cells.use,]
@@ -85,8 +98,8 @@ ModuleConnectivity <- function(
     )[genes_use,cells.use]
   }
 
+  # are we using the sparse correlation? (faster!)
   if(sparse){
-    
     kMEs <- corSparse(
       X = Matrix::t(exp_mat),
       Y = as.matrix(MEs)
@@ -103,7 +116,6 @@ ModuleConnectivity <- function(
       corOptions = corOptions,
       ...
     )
-
   }
 
   # add module color to the kMEs table
@@ -122,6 +134,45 @@ ModuleConnectivity <- function(
     seurat_obj <- ReassignModules(seurat_obj, harmonized=harmonized, wgcna_name=wgcna_name)
   }
 
+  # compute the intramodular degree for each gene
+  if('matrix' %in% class(TOM)){
+
+    degrees <- data.frame()
+    for(cur_mod in mods){
+
+      # get genes for this module
+      cur_genes <- modules %>% subset(module == cur_mod) %>% .$gene_name
+      cur_genes <- cur_genes[cur_genes %in% rownames(TOM)]
+
+      # subset TOM for genes in this module
+      cur_TOM <- TOM[cur_genes, cur_genes]
+      cur_TOM[lower.tri(cur_TOM)] <- 0 
+
+      # cast from from matrix to long dataframe, remove self-links
+      cur_TOM_df <- reshape2::melt(cur_TOM) %>% subset(Var1 != Var2)
+
+      # calculate degree for each gene and arrange by degree
+      cur_degrees <- cur_TOM_df %>% 
+        dplyr::group_by(Var1) %>% 
+        dplyr::summarise(degree = sum(value)) %>%
+        dplyr::mutate(weighted_degree = degree / max(degree)) %>%
+        dplyr::arrange(-degree) %>% 
+        dplyr::rename(gene_name = Var1)
+      
+      cur_degrees$module <- cur_mod 
+      degrees <- rbind(degrees, cur_degrees)
+    }
+    degrees <- degrees %>% dplyr::select(c(gene_name, module, degree, weighted_degree))
+    degrees <- as.data.frame(degrees)
+    degrees$module <- factor(as.character(degrees$module), levels=levels(modules$module))
+    degrees$gene_name <- as.character(degrees$gene_name)
+
+    # update the degree df in the Seurat object:
+    seurat_obj <- SetDegrees(seurat_obj, degrees, wgcna_name)
+
+}
+
+  # return seurat obj
   seurat_obj
 
 }